+++ models/db.js
@@ -1,19 +1,20 @@
-const mysql = require('mysql');
-const dbConfig = require("../config/db.config");
-//Create a connecttion to DB
-const connecttion = mysql.createConnection({
+const {Sequelize} = require("sequelize");
+const dbConfig = require("../config/db.config.js");
+
+//Create sequelize instanct ต้นแบบพิมพ์เขียว
+const sequelize = new Sequelize(dbConfig.DB , dbConfig.USER , dbConfig.PASSWORD,{
     host: dbConfig.HOST,
-    user: dbConfig.USER,
-    paaword: dbConfig.PASSWORD,
-    database:dbConfig.DB
-});
+    dialect:"mysql"
+})
 
-//Open Mysql
-connecttion.connect(
-    (error)=>{
-        if(error) throw error;
-        console.log("Seccessfully");
-    }
-);
-
-module.exports = connecttion;
+//Test Connect DB
+async function testConnection(){
+    try {
+        await sequelize.authenticate();
+        console.log('Connection has been established successfully.');
+        } catch (error) {
+        console.error('Unable to connect to the database:', error);
+      }
+}
+ testConnection();
+ module.exports = sequelize;
